refactor: Use different ID list for each list options
While bannertable json is still accessible, upcoming-shows json file is provided
with informations of shows currently not on air, but will be soon. So, combined
list of upcoming-shows and on-air replaces the roll of previous bannertable json
file.

When the list option is not supplied with 'live' argument, it is more reasonable
to print seperate output whether the show is on air.